Refactor Socket::select to be slightly less crazy (use array_map!)
I was unaware of this neat tool called array_map, and through reading excerpts of Adam Wathan's "Refactoring to Collections" - I became aware of it and realised that we needed a little refactoring here.

This might improve performance.  I'm not yet skilled enough to write the proper tests to check.